modify : controller modified

+++ server/controllers/locations/unsetLocations.js
@@ -1,19 +1,25 @@
 const { Location } = require("../../models");
 const { UserLocation } = require("../../models");
+const { isAuthorized } = require("../tokenFunctions");
 
 module.exports = async (req, res) => {
   try {
-    const { userId, location_name } = req.body;
+    const accessTokenData = isAuthorized(req);
+    const { location_name } = req.body;
 
-    const locationId = await Location.findOne({
-      where: { location_name: location_name },
-    });
+    if (!accessTokenData) {
+      return res.status(404).send({ message: "로그인이 되어있지 않습니다." });
+    } else {
+      const locationId = await Location.findOne({
+        where: { location_name: location_name },
+      });
 
-    await UserLocation.destroy({
-      where: { userId: userId, locationId: locationId.id },
-    });
+      await UserLocation.destroy({
+        where: { userId: accessTokenData.id, locationId: locationId.id },
+      });
 
-    res.status(200).json({ message: "즐겨찾기에서 삭제되었습니다." });
+      res.status(200).json({ message: "즐겨찾기에서 삭제되었습니다." });
+    }
   } catch {
     res.status(400).json({ message: "에러입니다." });
   }
